Start refactoring to clean up mess
